requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.104.1
h11==0.14.0
icalendar==5.0.11
idna==3.6
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
